feat(MOBILE-26):  enhance route functionality for open routes  in 2GIS and yandex maps
- Enhanced RouteModel with methods for generating Yandex and 2Gis URLs.
- Updated localization files to support new button labels for map services.
