--- lib/resources/app_icons.dart
@@ -1,7 +1,0 @@
-part of 'resources.dart';
-
-class AppIcons {
-  AppIcons._();
-
-  static const String cityIcon = 'assets/icons/city_icon.svg';
-}
