+++ lib/resources/app_images.dart
@@ -4,4 +4,6 @@ class AppImages {
   AppImages._();
 
   static const String bigTravellingCat = 'assets/images/big_travelling_cat.png';
+  static const String doubleGis = 'assets/images/double_gis.webp';
+  static const String yMaps = 'assets/images/y_maps.webp';
 }
