+++ lib/resources/app_vectors.dart
@@ -1,0 +1,7 @@
+part of 'resources.dart';
+
+class AppVectors {
+  AppVectors._();
+
+  static const String cityIcon = 'assets/vectors/city_icon.svg';
+}
