+++ lib/src/features/content/data/content_api_constants.dart
@@ -1,9 +1,9 @@
 abstract final class ContentApiConstants {
   // константы для реального сервера
-  static const port = 30080;
-  static const host = '109.196.98.55';
+  // static const port = 30080;
+  // static const host = '109.196.98.55';
 
   // константы для локального сервера
-  // static const port = 8087;
-  // static const host = '10.0.2.2';
+  static const port = 8087;
+  static const host = '10.0.2.2';
 }
