+++ spider.yaml
@@ -1,5 +1,5 @@
 # Generated by Spider
-# For more info on configuration, visit https://birjuvachhani.github.io/spider/grouping/  
+# For more info on configuration, visit https://birjuvachhani.github.io/spider/grouping/
 # Generates unit tests to verify that the assets exists in assets directory
 generate_tests: true
 
@@ -32,7 +32,7 @@ package: resources
 groups:
   - path: assets/images
     class_name: AppImages
-    types: [ .png, .jpg, .jpeg, .webp, .webm, .bmp ]
-  - path: assets/icons
-    class_name: AppIcons
-    types: [ .svg ]
+    types: [.png, .jpg, .jpeg, .webp, .webm, .bmp]
+  - path: assets/vectors/
+    class_name: AppVectors
+    types: [.svg]
