+++ test/app_images_test.dart
@@ -6,5 +6,7 @@ import 'package:flutter_travel_app/resources/resources.dart';
 void main() {
   test('app_images assets test', () {
     expect(File(AppImages.bigTravellingCat).existsSync(), isTrue);
+    expect(File(AppImages.doubleGis).existsSync(), isTrue);
+    expect(File(AppImages.yMaps).existsSync(), isTrue);
   });
 }
