+++ test/app_vectors_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_travel_app/resources/resources.dart';
 
 void main() {
-  test('app_icons assets test', () {
-    expect(File(AppIcons.cityIcon).existsSync(), isTrue);
+  test('app_vectors assets test', () {
+    expect(File(AppVectors.cityIcon).existsSync(), isTrue);
   });
 }
